diff reporter
https://coveord.atlassian.net/browse/CDX-1004
